requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
qrcode==7.4.2
Pillow==10.2.0 
flask-cors==4.0.0